♻️ slight refactoring, move setup_schema fn into entities::test module
