+++ crates/entities/src/models/crawl_queue.rs
@@ -280,9 +280,9 @@ mod test {
     use sea_orm::{ActiveModelTrait, Set};
     use url::Url;
 
-    use shared::config::{Limit, UserSettings};
     use crate::models::{crawl_queue, indexed_document};
     use crate::test::setup_test_db;
+    use shared::config::{Limit, UserSettings};
 
     #[tokio::test]
     async fn test_insert() {
