+++ crates/spyglass/src/api/route.rs
@@ -1,16 +1,16 @@
-use jsonrpc_core::{Error, ErrorCode, Result};
-use shared::response::LensResult;
 use entities::sea_orm::prelude::*;
 use entities::sea_orm::Set;
+use jsonrpc_core::{Error, ErrorCode, Result};
+use shared::response::LensResult;
 use tracing::instrument;
 use url::Url;
 
 use shared::request;
 use shared::response::{AppStatus, SearchLensesResp, SearchMeta, SearchResult, SearchResults};
 
+use entities::models::crawl_queue;
 use libspyglass::search::Searcher;
 use libspyglass::state::AppState;
-use entities::models::crawl_queue;
 
 use super::response;
